# api/specs/swagger.yaml
openapi: 3.0.0
info:
  title: Purchases API
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 0.1.9
servers:
  - url: http://localhost:8888/purchases
    description: Optional server description, e.g. Main (production) server
paths:
...
  /user:
    get:
      tags:
        - User
      operationId: getUsers
      summary: "Get list of all Users"
      responses:
        '200':
          description: "Retrieved list of all Purchases"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      tags:
        - User
      operationId: createUser
      summary: "Create user"
      requestBody:
        description: "User"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        '200':
          description: "Edited Purchase"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      tags:
        - User
      operationId: updateUser
      summary: "Edit user"
      requestBody:
        description: "User"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        '200':
          description: "Edited Purchase"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /user/password:
    put:
      tags:
        - User
      operationId: updatePassword
      summary: Update user password - disable with AD integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordUpdateHolder"
      responses:
        '200':
          description: "Updated password"
  /template:
    get:
      tags:
        - Purchase Template
      operationId: getTemplates
      summary: "Get list of all Templates"
      responses:
        '200':
          description: "Retrieved list of all Templates"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseTemplate"
    post:
      tags:
        - Purchase Template
      operationId: createTemplate
      summary: "Creates new PurchaseTemplate"
      requestBody:
        description: "Template"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseTemplate"
      responses:
        '200':
          description: "Created Purchase Template"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseTemplate"
    put:
      tags:
        - Purchase Template
      operationId: editTemplate
      summary: "Edit Purchase Template"
      requestBody:
        description: "Template"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseTemplate"
      responses:
        '200':
          description: "Edited Purchase Template"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseTemplate"
  /template/{templateId}:
    get:
      tags:
        - Purchase Template
      operationId: getTemplateById
      summary: "Get Purchase Template by id"
      parameters:
        - in: path
          required: true
          name: templateId
          description: "Id Purchase of the Template"
          schema:
            type: integer
      responses:
        '200':
          description: "Retrieved Purchase Template"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseTemplate"
    delete:
      tags:
        - Purchase Template
      operationId: deleteTemplate
      summary: "Deletes existing Purchase Template"
      parameters:
        - in: path
          required: true
          name: templateId
          description: "Id of the Purchase Template"
          schema:
            type: integer
      responses:
        '200':
          description: "Deleted Purchase Template"


# MODEL ----------------------------------------------------------------
components:
  schemas:
...
    PurchaseTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        text:
          type: string
        username:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseTemplateItem"
    PurchaseTemplateItem:
      type: object
      properties:
        id:
          type: integer
        purchaseTemplateId:
          type: integer
        itemName:
          type: string
        itemDescription:
          type: string
        ord:
          type: integer
...
    User:
      type: object
      properties:
        username:
          type: string
        fio:
          type: string
        apiKey:
          type: string
        gender:
          type: string
    PasswordUpdateHolder:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string